better writer checks
